Camera is now a component!!! :hamster:
Now that the camera is simplified into a component, it is updated right
after a game object's transform component and before any children which
saves some calls (compared to when it inherited GameObject).  Also moved
clearColor to camera instead of scene. This component thing is coming
together amazingly well!!! :cake: