made Config better
Now i can get a value based on an object.key string that I pass into it.
Example: "window.size.width" would go search the root object for window,
the window object for size, and the size object for the width key. I
would be returned the value of width. It works really nicely! Just a
little bit of ugly rapidjson code.